+++ vice/src/diskimage/fsimage.c
@@ -283,5 +283,5 @@ uint32_t fsimage_size(const disk_image_t *image)
     fsimage_t *fsimage;
 
     fsimage = image->media.fsimage;
-    return util_file_length(fsimage->fd);
+    return (uint32_t)util_file_length(fsimage->fd);
 }
